+++ src/main/java/org/ojalgo/structure/Access2D.java
@@ -127,7 +127,7 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
         default boolean isColumnSmall(final long row, final long col, final double comparedTo) {
             boolean retVal = true;
             final long tmpLimit = this.countRows();
-            for (long i = row; retVal && (i < tmpLimit); i++) {
+            for (long i = row; retVal && i < tmpLimit; i++) {
                 retVal &= this.isSmall(i, col, comparedTo);
             }
             return retVal;
@@ -150,7 +150,7 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
         default boolean isRowSmall(final long row, final long col, final double comparedTo) {
             boolean retVal = true;
             final long tmpLimit = this.countColumns();
-            for (long j = col; retVal && (j < tmpLimit); j++) {
+            for (long j = col; retVal && j < tmpLimit; j++) {
                 retVal &= this.isSmall(row, j, comparedTo);
             }
             return retVal;
@@ -247,9 +247,8 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
 
             if (delegateSpliterator != null) {
                 return new ElementView<>(delegateSpliterator, myStructure);
-            } else {
-                return null;
             }
+            return null;
         }
 
     }
@@ -275,7 +274,10 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
          * @deprecated v48 Will be removed
          */
         @Deprecated
-        long indexOfLargestInColumn(final long row, final long col);
+        default long indexOfLargestInColumn(final long row, final long col) {
+            long structure = this.countRows();
+            return this.indexOfLargestInRange(Structure2D.index(structure, row, col), Structure2D.index(structure, 0, col + 1));
+        }
 
         /**
          * @deprecated v48 Will be removed
@@ -411,8 +413,7 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
     }
 
     static boolean equals(final Access2D<?> accessA, final Access2D<?> accessB, final NumberContext accuracy) {
-        return (accessA.countRows() == accessB.countRows()) && (accessA.countColumns() == accessB.countColumns())
-                && Access1D.equals(accessA, accessB, accuracy);
+        return accessA.countRows() == accessB.countRows() && accessA.countColumns() == accessB.countColumns() && Access1D.equals(accessA, accessB, accuracy);
     }
 
     /**
@@ -430,13 +431,13 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
 
         if (anyElement instanceof ComplexNumber) {
 
-            for (int j = 0; retVal && (j < cols); j++) {
+            for (int j = 0; retVal && j < cols; j++) {
 
                 double imagDiag = ComplexNumber.valueOf(matrix.get(j, j)).i;
 
                 retVal &= PrimitiveScalar.isSmall(PrimitiveMath.ONE, imagDiag);
 
-                for (int i = j + 1; retVal && (i < rows); i++) {
+                for (int i = j + 1; retVal && i < rows; i++) {
 
                     ComplexNumber lowerLeft = ComplexNumber.valueOf(matrix.get(i, j)).conjugate();
                     ComplexNumber upperRight = ComplexNumber.valueOf(matrix.get(j, i));
@@ -450,8 +451,8 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
 
         } else {
 
-            for (int j = 0; retVal && (j < cols); j++) {
-                for (int i = j + 1; retVal && (i < rows); i++) {
+            for (int j = 0; retVal && j < cols; j++) {
+                for (int i = j + 1; retVal && i < rows; i++) {
 
                     double lowerLeft = matrix.doubleValue(i, j);
                     double upperRight = matrix.doubleValue(j, i);
@@ -515,7 +516,7 @@ public interface Access2D<N extends Comparable<N>> extends Structure2D, Access1D
         builder.append(matrix.getClass().getName());
         builder.append(' ').append('<').append(' ').append(numbRows).append(' ').append('x').append(' ').append(numbCols).append(' ').append('>');
 
-        if ((numbRows > 0) && (numbCols > 0) && (numbRows <= 50) && (numbCols <= 50) && ((numbRows * numbCols) <= 200)) {
+        if (numbRows > 0 && numbCols > 0 && numbRows <= 50 && numbCols <= 50 && numbRows * numbCols <= 200) {
 
             // First element
             builder.append("\n{ { ").append(matrix.get(0, 0));